add student_login and it has finish with 2 hour range only
