+++ app/views/layouts/application.html.erb
@@ -7,7 +7,13 @@
   <%= csrf_meta_tag %>
 </head>
 <body>
-  
+  <!-- trying my code (syamim) need to use flash -->
+    <div id="flash">
+      <% flash.each do |key, value| %>
+        <div class="alert alert-<%= key %>"><%= value %></div>
+      <% end %>
+    </div>
+
   <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
     <a class="navbar-brand" href="#">QRA</a>
       <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
@@ -19,10 +25,12 @@
           <li class="nav-item active">
             <a class="nav-link" href="#">Home <span class="sr-only">(current)</span></a>
           </li>
-          <% if current_user.teacher? || current_user.admin? %>
-          <li class="nav-item">
-            <a class="nav-link" href="/users/index">Dashboard</a>
-          </li>
+          <% if current_user %>
+            <% if current_user.teacher? || current_user.admin? %>
+            <li class="nav-item">
+              <a class="nav-link" href="/users/index">Dashboard</a>
+            </li>
+            <% end %>
           <% end -%>
           <li class="nav-item dropdown">
             <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
