+++ config/routes.rb
@@ -16,7 +16,7 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
   
   root 'home#index'
-  get "/auth/:provider/callback" => "sessions#create_from_omniauth"
+  get "/auth/:provider/callback" => "students#create_from_omniauth"
   get "/subjects" => "subjects#new", as: "subject"
   post "/subjects" => "subjects#create"
   get "/users/index"
